[UPDATE] 알림창 기능 재구현 및 알림아이콘기능 추가
