+++ src/app/notification/page.tsx
@@ -1,11 +1,10 @@
 import NotificationList from '@/components/notification/NotificationList';
-import Test from '@/components/notification/Test';
 import { Props } from '@/types/childrenPropsType';
 import React from 'react';
 
 function Notification() {
   // return <NotificationList />;
-  return <Test />;
+  return <NotificationList />;
 }
 
 export default Notification;
